Cleaned up code to better conform to standards, edited vim configs to help out with that.
Moved templated function definitions into *_impl.h

As for cout vs. printing; cout is better for primitive types but those variables
have to be typed, since there is no toString() method for all objects, the best
I can do is to store metadata about a variable's type and map that metadata to a
handler (which is what printing does) before passing it off to cout, after 
properly typecasting it.

That said, ripping out or significantly modifying this feature will be trivial 
due to prior refactoring of optionshierarchy.