Be sure to serialize and deserialize in the same way!
In this case I serialized as a pointer and deserialized as an object.  This
causes disaster, but figuring out what the disaster was required digging into
Boost, recompiling with debugging symbols, stepping through the nightmare
serialization code, reasoning about what was going wrong, and ultimately fixing
it by changing one single character.  Sigh.  Let this be a warning to whomever
reads this, although that person will probably only read this after having
already gone through the horrible process outlined above...