Force a minimum value on the diagonal.
This is necessary so that the inverse covariance doesn't explode.  If the
covariance is sufficiently small, then the PDF of the Gaussian will condense
towards a delta function, and you will end up with inf's in your probabilities,
which turn into NaNs when normalized, and then your entire model turns into
NaNs.

Note that the PositiveDefinite constraint can be replaced with the NoConstraint
class for speed if necessary.