Fix bug where arma_config.hpp wouldn't be installed.
This would happen when the user configured once and immediately installed.  The
issue was that MLPACK_SRCS had arma_config.hpp appended to it too late, and
didn't exist at the time of the first scan of files.  When CMake was run a
second time, arma_config.hpp existed and thus would get picked up by the scan
and added to MLPACK_SRCS.  This is a bug I didn't run into because I don't build
from a new environment every time...