Fix dual-tree k-means runtime bug.
The residue was being calculated incorrectly, and as a result it would run until
the maximum number of iterations.  Thanks to Erich Schubert for bringing this
one up and pointing it out.

The fascinating thing is that this did not trigger any test failures, because it
still converged to the correct results---it just took way longer than it should
have.