Fix unusual bug that pruned too tightly.
It is possible, in very weird and occasional situations (in this case, by using
mlpack_allkfn with phy.csv and cover trees), that a point in a candidate
reference node would have the exact same distance as the bound distance.  But in
this case the reference node should not be pruned, since that point in the
candidate reference node can still update the results!  That could happen if,
for instance, the results array for a particular point isn't completely full
yet.