Fix more subtle memory bugs.
First: don't transfer the memory state to Armadillo.  Instead, make a non-strict
alias; this means that Armadillo will only allocate new memory on a size change,
and we won't end up making the Python user's numpy array be nothing when we're
done because we std::move()d the memory.

Second: Only transfer the ownership of the matrix to numpy if the Armadillo
matrix actually allocated its own memory.  This can help with situations where
the user passes a numpy matrix that they have a reference to that will also be a
part of the output.