Use iterate dimension for particle shape and set it as the starting point
This commit updates the PSO algorithm and uses the iterate dimension for representing a particle. Futhermore, arma::cube is used to represent the entire population, however, this is resource expensive, so it has to be optimized later on. The test case is also updated to use a more accurate starting point.