txt2man options
Pass appropriate options to txt2man to get a proper .TH line.
These are -s 1 for the section,
we skip -d xxx for the date, which will default to the
debian/changelog timestamp during a debian package build, or otherwise
to the current date.
Then -r origin-aka-project-name-and-version -v volume-aka-section-name.

Gbp-Pq: Name 0003-txt2man-options.patch