Update tests for changed APIs.
Random forests generally work better, but it is not guaranteed for the vc2
dataset, so I am still requiring only that it gets 90% of the decision tree
performance in the worst case.  I expect it will generally be better, but there
are still situations where it may not be (because of the randomness).