Add Azure (continuous integration) pipelines config file.
Up to this point, we used a combination of Travis CI (Linux and macOS) and
Appveyor (Windows) to test each PR. Recently we encountered some timeouts on the
Travis side as building and testing mlpack can take some time; a timeout is
marked as failure which can be confusing, especially for new contributors. In
the past we asked the Travis/Appveyor teams to bump the timeout for the mlpack
project, but with Azure Pipelines which offers hosted pipelines for Linux,
macOS, and Windows with 10 parallel jobs and 360 minutes of compute for open
source projects I think there is no need to.