Add path for all each boost dependency
It seems that cmake requires to add the path to each library when these
library used as dependency. CMake is not capable of finding these libraries.

I have tried several ways to look for boost. Since CMakeLists uses
find_package() as the standard method of searching boost. It is a good idea to
find the path for boostConfig.cmake using the PATHS variable inside
find_package(). However, this method does not work and CMake used to look for
boost installed in my /usr/lib/.

Another method is to change the CMAKE_PREFIX_PATH. This path is used by CMake to
look for dependencies at a specific location. After several trial, it gave the
same result as PATHS in find_package().

It seems very complicated for cmake to look for dependencies, I am not able to
figure the reason behind this. The only possible solution is to give the path
for each dependency manually when starting cmake ../

Finally, I had to download CLAPACK instead of lapack and link against it. Since
LAPACK uses Fortran, we need to cross-compile fortran code and link agaist main
fortran library.

I linked Armadillo against OpenBLAS CLAPACK and CBLAS. I am not sure in this
case if OpenBLAS is being used, more investigation is needed. The linker
complains of undefined references used by CLAPACK if CBLAS is not used.

More work will be done here. So far, mlpack_knn is compiling and running on
my RPI :)

Signed-off-by: Omar Shrit <shrit@lri.fr>