Protect Armadillo from getting error if static link
If `ARMA_USE_WRAPPER` is `ON`, this patch will allow to find armadillo.a
if we are trying to statically link, it armadillo is not found then it
should stop there.

If we are not building statically and `ARMA_USE_WRAPPER` is ON, then
everything should be fine.

If we are cross compiling, the we are not using `ARMA_USE_WRAPPER`

Signed-off-by: Omar Shrit <omar@shrit.me>