Remove ar parameter to avoid the warning
Serialization is not used in this layer, therefore the warning of
unused parameter is becoming annoying when trying to compile the
mlpack_test making it impossible to find different errors in small
terminal. Since we are using boost_visitor, the unfolding of layers
is taking several pages of terminal to print this small warning.

Hoping that this will mitigate one of the warning.

Signed-off-by: Omar Shrit <omar@shrit.me>