Fix compilation errors with clang 19
The new compiler version seems to perform more exhaustive checks on
templates, even if they are not initiated. As the code was malformed
also before, it is likely that it is dead code. However, this commit is
mainly to ensure users using clang 19 can include the header without
compilation error.

Note that this commit does not guarantee that currently other code might
not compile with clang 19, it just fixes some code which was used
transitevily and caused problems.