Add script to format code base
In order to guarantee a standardized formatting this commit prepares a
`.clang-format` file which tries to follow the previously used coding
conventions (i.e. trying to cause a minimal diff without too extensive
fine-tuning). The formatting is checked as part of the CI.

Note that the commit focuses on formatting without the include
directives, which shall be tackled in a dedicated commit.