Modernize STL type traits
With newer standards the readability of template heavy code can
be improved a lot by using a form with trailing _t for ::type and
_v for ::value respectively. Especially for ::type it also makes
the additional typename before the trait dispensable.

This PR combines multiple commits that modernize the traits,
starting with what clang-tidy could achieve while analyzing test files
and continuing by regex replaces with few manual corrections.

While it does not guarantee that all current occurrences are modernized,
it moves the codebase considerably into the modern direction. Note that
also before this PR both notations were used, so missing a few should
be acceptable.