Fix armadillo deprecation warning
The use of `operator<<` to fill matrices is deprecated since a little
more than two years. In order to reduce the number of warnings during
compilation of `mlpack` and reduce likelyhood that something breaks as
soon as `armadillo` removes the deprecated operator, this commit uses
the initializer list form as suggested by the deprecation warning from
`armadillo`.